TCF Agent: fixed: malloc trace does nor work on Windows - MSVC does not support printf format specifier %zd

+++ framework/myalloc.c
@@ -34,7 +34,7 @@ void * loc_alloc(size_t size) {
         perror("malloc");
         exit(1);
     }
-    trace(LOG_ALLOC, "loc_alloc(%zd) = %#lx", size, p);
+    trace(LOG_ALLOC, "loc_alloc(%u) = %#lx", (unsigned)size, p);
     return p;
 }
 
@@ -51,7 +51,7 @@ void * loc_alloc_zero(size_t size) {
         exit(1);
     }
     memset(p, 0, size);
-    trace(LOG_ALLOC, "loc_alloc_zero(%zd) = %#lx", size, p);
+    trace(LOG_ALLOC, "loc_alloc_zero(%u) = %#lx", (unsigned)size, p);
     return p;
 }
 
@@ -67,7 +67,7 @@ void * loc_realloc(void * ptr, size_t size) {
         perror("realloc");
         exit(1);
     }
-    trace(LOG_ALLOC, "loc_realloc(%#lx, %zd) = %#lx", ptr, size,p);
+    trace(LOG_ALLOC, "loc_realloc(%#lx, %u) = %#lx", ptr, (unsigned)size, p);
     return p;
 }
